+++ src/lib/react-hook-form/ImagePreview/index.tsx
